move logic to extension functions and generalize
